# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage news and courses.",
        "title": "News and Courses API",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/articles": {
            "post": {
                "description": "创建一个新的文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "创建新文章",
                "parameters": [
                    {
                        "description": "文章信息",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "description": "根据文章ID获取文章详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses": {
            "get": {
                "description": "获取所有课程",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "获取课程列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Course"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的课程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "创建新课程",
                "parameters": [
                    {
                        "description": "课程信息",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{id}": {
            "get": {
                "description": "根据课程ID获取课程详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "获取课程详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "课程ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "description": "获取新闻列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "获取新闻列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "新闻来源",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.News"
                            }
                        }
                    }
                }
            }
        },
        "/news/refresh": {
            "post": {
                "description": "强制刷新新闻数据",
                "tags": [
                    "news"
                ],
                "summary": "刷新新闻列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Article": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Article Content",
                    "type": "string"
                },
                "course_id": {
                    "description": "Foreign key to Course",
                    "type": "integer"
                },
                "created_at": {
                    "description": "Created Time",
                    "type": "string"
                },
                "id": {
                    "description": "Unsigned bigint",
                    "type": "integer"
                },
                "title": {
                    "description": "Article Title",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Updated Time",
                    "type": "string"
                }
            }
        },
        "model.Course": {
            "type": "object",
            "properties": {
                "articles": {
                    "description": "One-to-many relationship",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Article"
                    }
                },
                "created_at": {
                    "description": "Created Time",
                    "type": "string"
                },
                "description": {
                    "description": "Course Description",
                    "type": "string"
                },
                "id": {
                    "description": "Unsigned bigint",
                    "type": "integer"
                },
                "title": {
                    "description": "Unique Title",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Updated Time",
                    "type": "string"
                }
            }
        },
        "model.News": {
            "description": "新闻",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "source": {
                    "description": "Source means different news sources, such as Weibo, Zhihu, etc.",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}